Task 1: Correctness fixes (5 fixes)
Fix 1: Decrement suggestion frequency on removeDoc()
- Add docWords tracking map (docId -> word -> count)
- On removeDoc, decrement frequency per word and prune when zero

Fix 2: Throw error on mixed AND/OR filter groups
- Detect when both AND and OR are specified in same filter group
- Error message: 'AND and OR are mutually exclusive...'

Fix 3: Handle circular references in flattenValue()
- Add visited Set to detect and skip circular references
- Gracefully return empty string for already-visited objects

Fix 4: Validate config on import() with warning
- Compare serialized config fields against provided config
- Log warning on mismatch, document idResolver requirement

Fix 5: Uint16Array -> Uint32Array in levenshtein
- Prevent silent overflow for strings > 65535 chars

+++ src/core/levenshtein.ts
@@ -23,8 +23,8 @@ export function levenshtein(a: string, b: string, maxDist = 2): number {
   if (lb === 0) return la <= maxDist ? la : Infinity;
 
   // Two-row DP
-  let prev = new Uint16Array(lb + 1);
-  let curr = new Uint16Array(lb + 1);
+  let prev = new Uint32Array(lb + 1);
+  let curr = new Uint32Array(lb + 1);
 
   for (let j = 0; j <= lb; j++) prev[j] = j;
